feat: impl for remaining needed gitlab s3 ops (#56)
* feat: impl for deleteObjects

* fix: swift bulk-delete issue with deno fetch

* feat: ListBuckets op

* fix: AbortMultipartUpload not deleting parts

* chore: polish abourt MPU logging

* feat: ranged GetObject

* chore: update ranged get object test

* chore: add new error type

* fix: list parts prefixing object key into search params

* fix: request routing

* fix: root url handler

+++ src/backends/s3/mod.ts
@@ -4,6 +4,7 @@ import {
   copyObject,
   createMultipartUpload,
   deleteObject,
+  deleteObjects,
   getObject,
   headObject,
   listObjects,
@@ -13,15 +14,14 @@ import {
   createBucket,
   deleteBucket,
   headBucket,
+  listBuckets,
   routeQueryParamedRequest,
 } from "./buckets.ts";
-import { HeraldError } from "../../types/http-exception.ts";
 import { areQueryParamsSupported } from "../../utils/url.ts";
 import { extractRequestInfo } from "../../utils/s3.ts";
 import { Bucket } from "../../buckets/mod.ts";
 import { RequestContext } from "../../types/mod.ts";
-import { APIErrors, getAPIErrorResponse } from "../../types/api_errors.ts";
-import { createErr, createOk, Result } from "option-t/plain_result";
+import { createOk, Result } from "option-t/plain_result";
 
 const handlers = {
   putObject,
@@ -37,6 +37,8 @@ const handlers = {
   createMultipartUpload,
   completeMultipartUpload,
   abortMultipartUpload,
+  deleteObjects,
+  listBuckets,
 };
 
 export async function s3Resolver(
@@ -47,7 +49,10 @@ export async function s3Resolver(
   const logger = reqCtx.logger;
 
   // FIXME: `resolveHandler` has already extracted request info
-  const { method, objectKey, queryParams } = extractRequestInfo(request);
+  const { bucket, method, objectKey, queryParams } = extractRequestInfo(
+    request,
+  );
+  const url = new URL(request.url);
   const queryParamKeys = new Set(Object.keys(queryParams));
 
   logger.debug(`Resolving S3 Handler for Request...`);
@@ -60,21 +65,29 @@ export async function s3Resolver(
         return await handlers.listObjects(reqCtx, request, bucketConfig);
       }
 
-      if (!areQueryParamsSupported(queryParamKeys)) {
-        logger.critical("Unsupported Query Parameter Used");
-        return createErr(
-          new HeraldError(400, {
-            message: "Unsupported Query Parameter Used",
-          }),
+      if (areQueryParamsSupported(queryParamKeys)) {
+        return await handlers.routeQueryParamedRequest(
+          reqCtx,
+          request,
+          bucketConfig,
+          queryParamKeys,
         );
       }
-      return await handlers.routeQueryParamedRequest(
-        reqCtx,
-        request,
-        bucketConfig,
-        queryParamKeys,
-      );
+
+      if (url.pathname === "/") {
+        return await handlers.listBuckets(reqCtx, request, bucketConfig);
+      }
+
+      break;
     case "POST":
+      if (queryParamKeys.has("delete")) {
+        return await handlers.deleteObjects(
+          reqCtx,
+          request,
+          bucketConfig,
+        );
+      }
+
       if (objectKey && queryParamKeys.has("uploads")) {
         return handlers.createMultipartUpload(reqCtx, request, bucketConfig);
       }
@@ -87,11 +100,7 @@ export async function s3Resolver(
         );
       }
 
-      return createErr(
-        new HeraldError(403, {
-          message: "Unsupported request",
-        }),
-      );
+      break;
     case "PUT":
       if (objectKey && request.headers.get("x-amz-copy-source")) {
         return await handlers.copyObject(reqCtx, request, bucketConfig);
@@ -101,7 +110,11 @@ export async function s3Resolver(
         return await handlers.putObject(reqCtx, request, bucketConfig);
       }
 
-      return await handlers.createBucket(reqCtx, request, bucketConfig);
+      if (bucket) {
+        return await handlers.createBucket(reqCtx, request, bucketConfig);
+      }
+
+      break;
     case "DELETE":
       if (objectKey && queryParamKeys.has("uploadId")) {
         return await handlers.abortMultipartUpload(
@@ -114,14 +127,28 @@ export async function s3Resolver(
         return await handlers.deleteObject(reqCtx, request, bucketConfig);
       }
 
-      return await handlers.deleteBucket(reqCtx, request, bucketConfig);
+      if (bucket) {
+        return await handlers.deleteBucket(reqCtx, request, bucketConfig);
+      }
+
+      break;
     case "HEAD":
       if (objectKey) {
         return await handlers.headObject(reqCtx, request, bucketConfig);
       }
-      return await handlers.headBucket(reqCtx, request, bucketConfig);
+
+      if (bucket) {
+        return await handlers.headBucket(reqCtx, request, bucketConfig);
+      }
+
+      break;
     default:
-      logger.critical(`Unsupported Request Method: ${method}`);
-      return createOk(getAPIErrorResponse(APIErrors.ErrInvalidRequest));
+      logger.warn(
+        `Unsupported Request Method: ${method} on ${request.url}`,
+      );
+      return createOk(new Response("Proxy is running..."));
   }
+
+  logger.warn(`Unsupported Request: method ${method} on ${request.url}`);
+  return createOk(new Response("Proxy is running..."));
 }
